fix(sdk): rename quick_client to quic_client
